🐛When we are creating path for the files to store them in S3, sometimes it's with trailing "/" and sometimes without
